🚚 feat(graph_utils): move graph utils from `graphs.py` to `graph_utils.py`
